// src/api/generated/models/leadSource.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

/**
 * @nullable
 */
export type LeadSource = typeof LeadSource[keyof typeof LeadSource] | null;


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LeadSource = {
  portal: 'portal',
  site: 'site',
  phone: 'phone',
  referral: 'referral',
  social_media: 'social_media',
  walk_in: 'walk_in',
  other: 'other',
} as const;