// src/api/generated/models/leadStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LeadStatus = {
  new: 'new',
  contacted: 'contacted',
  negotiating: 'negotiating',
  converted: 'converted',
  lost: 'lost',
} as const;